requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
albumentations
Pillow
tqdm
prettytable